services/web-app/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
redis==5.0.7
pika==1.3.2
python-dotenv==1.1.1
boto3==1.39.11
flasgger==0.9.7.1
gunicorn==23.0.0
Faker==26.0.0
tqdm==4.66.4
pytest==8.3.2
pytest-cov==6.2.1
pytest-mock
Flask-SocketIO==5.3.7
gevent==25.5.1
gevent-websocket==0.10.1
line-bot-sdk==3.17.1
mutagen==1.47.0
apscheduler==3.10.4
Flask-APScheduler==1.13.1
# 向量資料庫和 AI 相關
pymilvus==2.3.7
openai==1.35.0  # 使用更新的穩定版本
marshmallow==3.20.1  # 修復 pymilvus 依賴問題
environs==9.5.0  # 符合 pymilvus 要求
httpx==0.25.0  # 指定 httpx 版本

# 資料處理（使用預編譯版本）
numpy==1.24.3  # 固定 numpy 版本以避免衝突
pandas==2.0.3
openpyxl==3.1.2

# 確保 grpcio 使用預編譯版本
grpcio>=1.57.0,<1.60.0
